cc: Generate tilings at other scales for impl-side painting
This patch adds infrastructure to make it easy for PictureLayerImpl to
manage its set of tilings via PictureLayerImpl::ManageTiles().

The main functionality change is to additionally generate a low-res set
of tiles (by default 1/8 resolution).  Along with generating tilings at
different scales, tile priorities now have the correct resolution set on
them so that they can be prioritized correctly.

There's a few small bugfixes added here as well for correctness, such as
making sure that tilings are sorted in the PictureLayerTilingSet from
highest to lowest.  (This is the order in which tiles will be used.)  It
also correctly calculates a synthetic contents scale / content bounds for
the PictureLayerImpl to support multiple active contents scales.

R=nduca@chromium.org
BUG=167300


Review URL: https://chromiumcodereview.appspot.com/11704002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@175927 0039d316-1c4b-4281-b951-d872f2087c98