Improve thread safety for device changes on OSX.
Device changes are hard, lets go shopping!  Sadly OSX does not handle property
listener callbacks in a thread safe manner.  On 10.6 we can set
kAudioHardwarePropertyRunLoop to account for this.  On 10.7 this is broken
and we need to create a dispatch queue to drive callbacks.  However code
running on the dispatch queue must be mutually exclusive with code running
on the audio thread.

ExclusiveDispatchQueueTaskObserver works around this by pausing and resuming
the dispatch queue before and after each pumped task.  Benchmarks on my retina
10.8.2 macbook pro show on average this takes < 30us in total.  No performance
issues were seen even using FakeAudioOutputStream which drives audio callbacks
on the message loop.

Complicating matters it the fact that the AudioProperty(Add|Remove)ListenerBlock
functions are not available on 10.6... so we need to dynamically load them :-/

This is not ideal and long term we should replace the audio thread on OSX with a
dispatch queue to avoid having to suspend/resume the queue; this is non-trivial
and requires a significant refactoring of the AudioManager to support.

BUG=158170
TEST=media_unittests. manual device changes.

Review URL: https://codereview.chromium.org/11824018

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176228 0039d316-1c4b-4281-b951-d872f2087c98