Enables compositing support for webview.
Mailboxes for the webview are received from RWHVGuest and set on the
texture layer with a callback that gets triggered when they are no
longer in use by the compositor.

An IPC is sent back to the browser with a sync point to ack
the presented buffer.

BUG=143429


Review URL: https://chromiumcodereview.appspot.com/11824040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176310 0039d316-1c4b-4281-b951-d872f2087c98