Inconsistent default zoom handling in incognito mode
The HostZoomMap(Impl) is normally initialized through the 
User Profile but when in Incognito a the newly created
HostZoomMap(Impl) is initialized by copying the HostZoomMap(Impl)
of the parent non-Incognito version.

The problem is that HostZoomMapImpl::CopyFrom only copies the
content of the map but not the default_zoom_level_.  This results
in new Incognito pages using a zoom level of zero which is the
constructor initialized value of default_zoom_level_.

Note: I do not have commit rights.
Contributed by tom.cassiotis@gmail.com

BUG=168989
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/11878005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176587 0039d316-1c4b-4281-b951-d872f2087c98