Test that 404 subresource bodies with nested message loops don't crash.
This is a browser test because DumpRenderTree doesn't run nested message loops. The failure case was that a nested message triggered from an element that has signalled an error but had an open request would receive a body for the request and crash/fail an assertion.

This test checks the WebKit side fix for this issue, see https://bugs.webkit.org/show_bug.cgi?id=103563

R=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11778083

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176882 0039d316-1c4b-4281-b951-d872f2087c98