Add unsupported option to install-build-deps.
Add an option --unsupported to install-build-deps.sh which allows the
installation to be attempted on a system we don't support. Omit
mentioning this in the error, in the hope that anyone who can deal with
the fallout from this will find it in the script. Also, add GCEL
(Ubuntu-derived image used by Compute Engine) to the list of supported
systems.

BUG=

Review URL: https://codereview.chromium.org/11863012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@176982 0039d316-1c4b-4281-b951-d872f2087c98