Fix for a crasher when starting a renderer.
A crash dump from the field reports that the CrashDumpManager in
ChromeContentBrowserClient is not initialized when
GetAdditionalMappedFilesForChildProcess is called.
This is because the CrashDumpManager is initialized when a RenderProcessHost
is created, but it could be the case that a BrowserChildProcessHost gets
created first and calls GetAdditionalMappedFilesForChildProcess before
any RenderProcessHost was created.
Making the CrashDumpManager a singleton initialized early on to prevent this
crasher.

BUG=163456
TEST=None


Review URL: https://chromiumcodereview.appspot.com/11884042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177110 0039d316-1c4b-4281-b951-d872f2087c98