Add a helper for querying connectivity state
This is the first pass at a basic connectivity helper which lets consumers
query the connectivity state on Chrome OS. Depending on the value of the
--enable-new-network-handlers flag, the helper will return this information
from the network library or the new network state handler.

Right now it supports querying for a connecting/connected network of a 
given type, and returning the name of the first network of a given type 
(according to priority order). 

Also, changed the login helper to use this helper instead of directly
accessing network library.

BUG=chromium:162716

For chrome/chrome_browser_chromeos.gypi OWNERS
TBR=sky@chromium.org 

Review URL: https://chromiumcodereview.appspot.com/11642058

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177366 0039d316-1c4b-4281-b951-d872f2087c98