cc: Fix PictureLayerTiling::Iterator crash
If you gave PictureLayerTiling a rect that was outside of its tiling
data, then it would incorrectly return its last tile as if that was
valid.  Fixed with an early-out and a unit test.

This situation should never have occurred however, as the bounds for the
tiling data should be absolutely equal to the layer's bounds.  It turned
out that during tiling sync, the active layer bounds were being set on
the pending layer tiling data instead of rewriting it with the new bounds.

R=danakj@chromium.org
BUG=170768


Review URL: https://chromiumcodereview.appspot.com/12026006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177714 0039d316-1c4b-4281-b951-d872f2087c98