Refactor FileIO to the new resource host system. Taking over from Victor's CL: https://codereview.chromium.org/11419131/
Re-landing, with a fix. The original waited to reset the file state until
after the callback had completed. This is too late, as the client should be
able to perform another file operation during the callback.

Changeset #1 is the original patch. The fix is in change set #2

Original author=Victor Hsieh
BUG=none
TEST=browser_tests --gtest_filter=PPAPINaClNewlibTest.FileIO*

Review URL: https://chromiumcodereview.appspot.com/11941022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177830 0039d316-1c4b-4281-b951-d872f2087c98