Adding proof-of-concept WebRTC browser test.
This test is imagined as the successor to the WebRTC PyAuto tests. It will require real devices and will be disabled in the main test suite (it will run on special bots we set up though).

The test uses the same javascript as the PyAuto tests. One problem is that the PyAuto tests use WaitUntil polling to get results from some javascript events. I would like to rewrite that for this test so we use the DOMMessageQueue instead, but I can't do that yet since that would break our existing, running tests. That will have to wait until we have ported all tests.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11886042

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@177917 0039d316-1c4b-4281-b951-d872f2087c98