[android] Fix size reported in contentSizeChange callback.
This makes the onContentSizeChanged callback take into account
page scale when returning content size.
This is required because the width and height values provided
by the renderer are in a different coordinate space and need
to be divided by the page scale factor to get "screen pixel"
values.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/12042017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@178039 0039d316-1c4b-4281-b951-d872f2087c98