cc: Give total impl-side page scale to calcDrawProps.
When we are in a pinch on poster circle, the scale of the layers falls back to
the page scale factor, since their combined transform is not a simple
translate/scale.

We pass in the page_scale_factor which does not include the impl-side delta.
This means we are giving the main-thread page scale to calcDrawProps, so we
don't raster the layers with the appropriate scale.

R=enne
BUG=172357


Review URL: https://chromiumcodereview.appspot.com/12082004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@178994 0039d316-1c4b-4281-b951-d872f2087c98