History: Fix DCHECK triggered by some unnecessary code.
When advancing to the second page of history in a debug build, a
DCHECK is triggered in history_types.cc, caused by passing both
|end_time| and |cursor| in the QueryOptions.

The cursor actually eliminates the need to pass |end_time| at all
when paging through history, so it's safe to remove that code
altogether.

BUG=None
TEST=Run a debug build with >150 history entries, open chrome://history,
click the "Older" link at the bottom, and verify that the DCHECK is not
hit.


Review URL: https://chromiumcodereview.appspot.com/12086019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179203 0039d316-1c4b-4281-b951-d872f2087c98