chromeos: Support delaying suspend asynchronously.
This adds PowerManagerClient::GetSuspendReadinessCallback(),
which PowerManagerClient::Observer implementations can call
from within SuspendImminent().  The method returns a Closure
that the observer can call after doing asynchronous work to
report that it's ready for the system to be suspended.

It also makes SessionManagerClient listen for ScreenIsLocked
and ScreenIsUnlocked signals and removes an unused
GetIsScreenLocked() method.

BUG=171960,128798

Review URL: https://codereview.chromium.org/12095006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179350 0039d316-1c4b-4281-b951-d872f2087c98