Reverted the appearance of simple notifications.
A certain amount of existing app, extension, and OS code uses the
existing webkitNotifications API to generate simple notifications.
http://crrev.com/11926032 changed the appearance of those notification
in a way that causes some of that existing code to now generate
notifications with pixelated icons and clipped text.

This change reverts the look of just simple notifications to avoid this.
The new look can still be obtained with either of the flags
--enable-rich-notifications (available through chrome://flags, but
currently has some side effects) or --enable-new-simple-notifications
(not available through chrome://flags, but no side effects).

BUG=172358


Review URL: https://chromiumcodereview.appspot.com/12090011

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179423 0039d316-1c4b-4281-b951-d872f2087c98