Delete GPU channel on failure to reallocate backbuffer.
This will cause a lost context and the renderer process will attempt to recover. If it does not recover, it will fall back to software compositing.

Currently SetBackbufferAllocation can only fail on Windows but we should detect failures on other platforms as well.

This might potentially fix the referenced bug.

BUG=170767

Review URL: https://chromiumcodereview.appspot.com/12090027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179471 0039d316-1c4b-4281-b951-d872f2087c98