chromeos: Lock before suspending when requested.
When the system is about to suspend, make Chrome request
that the screen be locked if the lock-on-suspend pref is
set.  Chrome then delays the suspend attempt until it
receives notice that the lock screen is visible.  powerd was
previously responsible for doing this.

BUG=171960,chromium:128798,chromium-os:38152,chromium-os:36889,chromium-os:38281


Review URL: https://chromiumcodereview.appspot.com/12092050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179651 0039d316-1c4b-4281-b951-d872f2087c98