Browser Plugin: Software path failover no longer crashes the embedder.
When failing over to the software rendering path from the compositing path,
BrowserPlugin would use the old backing store that was created prior to
switching to the compositing path. If the BrowserPlugin was resized while
compositing, then the backing store would be rendered incorrectly and 
BrowserPlugin might attempt to read/write beyond the end of the buffer and crash.

BUG=171978
Test=Manually killing the GPU process three times while viewing Poster Circle


Review URL: https://chromiumcodereview.appspot.com/12079050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@179925 0039d316-1c4b-4281-b951-d872f2087c98