cc: Fix the PictureLayerImplTest.addTilesFromNewRecording test.
The test had 0 tilings since the recorded region was empty. And it
was not using the full tile rect when deciding if the tile would be
present or not, which does not match the code in the CreateTile() logic
so the test would fail.

R=enne
BUG=173782
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12090126

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@180231 0039d316-1c4b-4281-b951-d872f2087c98