Temporary workaround to un-break WebView rendering.
This ignores the size provided to SetSize, which can be in CSS pixels from WebContentsImpl, to always use the physical size from ContentViewCore. However, this value should probably not be ignored. This should be a temporary fix to unblock WebView until a proper solution is worked out.

BUG=174100


Review URL: https://chromiumcodereview.appspot.com/12217012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@180818 0039d316-1c4b-4281-b951-d872f2087c98