overscroll: Use wheel events only from the trackpad.
Using wheel events from an external mouse is possible to use for overscroll,
in theory. But, the problem with that is we currently do not receive any event
that indicates that scrolling has ended. For trackpad, on the otherhand, we do
receive a fling event (with 0 velocity) at the end of a 2-finger scroll. So
do overscrolling from trackpad, but not from normal mouse wheels.

BUG=174508

Review URL: https://codereview.chromium.org/12226025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@180973 0039d316-1c4b-4281-b951-d872f2087c98