Add GetUploadStatus method to the DriveServiceInterface.
This CL adds a (stub) method GetUploadStatus to the DriveServiceInterface.
The goal of adding this method is supporting retry on receiving 5xx http
response from server for data uploading.
The callback type of the method is as same as the one for ResumeUpload.
So, it also renames ResumeUploadResponse and ResumeUploadCallback to
UploadRangeResponse and UploadRangeCallback respectively.
The params of the method is based on the experimental implementation.
The actual implementation of {gdata_wapi,drive_api}_{service,operations} should
be done in following CLs.

BUG=148632
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/12235002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@181245 0039d316-1c4b-4281-b951-d872f2087c98